1.模拟IIC读取陀螺仪数据实现，用的是PB4,PB5需禁止JTUG的功能才能为普通IO口。 2.系统滴答定时器使用。

+++ USER/stm32f10x_it.c
@@ -79,7 +79,7 @@ void PendSV_Handler(void)
 u8 Flag_1ms=0,Flag_5ms=0,Flag_10ms=0,Flag_20ms=0,Flag_100ms=0,Flag_500ms=0,Flag_1000ms=0;
 void SysTick_Handler(void)
 {
-		static u16 Time_cnt=0;
+		static u32 Time_cnt=0;
 		Time_cnt++;
 		Flag_1ms=1;
 		if(Time_cnt%10==0)
